decompose exhaustive deps rule helpers
